Refactor, add window state
